+++ apps/builder/app/builder/features/sidebar-left/panels/marketplace/marketplace.tsx
@@ -12,15 +12,14 @@ import type { MarketplaceRouter } from "~/shared/marketplace/router";
 import { createTrpcFetchProxy } from "~/shared/remix/trpc-remix-proxy";
 import type { Project } from "@webstudio-is/project";
 import { SettingsPanel } from "../pages/settings-panel";
-import { ItemDetails } from "./item-details";
+import { About } from "./about";
 
 const trpc = createTrpcFetchProxy<MarketplaceRouter>(marketplacePath);
 
 export const TabContent = ({ onSetActiveTab }: TabContentProps) => {
   const [activeOverviewItem, setAciveOverviewItem] =
     useState<MarketplaceOverviewItem>();
-  const [openDetailsProjectId, setOpenDetailsProjectId] =
-    useState<Project["id"]>();
+  const [openAbout, setopenAbout] = useState<Project["id"]>();
 
   const {
     load: getItems,
@@ -34,9 +33,7 @@ export const TabContent = ({ onSetActiveTab }: TabContentProps) => {
     getItems();
   }, [getItems]);
 
-  const openDetailsItem = items?.find(
-    (item) => item.projectId === openDetailsProjectId
-  );
+  const openAboutItem = items?.find((item) => item.projectId === openAbout);
 
   return (
     <Flex direction="column" css={{ height: "100%" }}>
@@ -67,16 +64,16 @@ export const TabContent = ({ onSetActiveTab }: TabContentProps) => {
             setAciveOverviewItem(activeOverviewItem);
             getBuildData({ projectId: activeOverviewItem.projectId });
           }}
-          openDetailsProjectId={openDetailsProjectId}
-          onOpenDetailsProjectIdChange={setOpenDetailsProjectId}
+          openAbout={openAbout}
+          onOpenAbout={setopenAbout}
         />
       )}
-      {openDetailsItem && (
+      {openAboutItem && (
         <SettingsPanel isOpen>
-          <ItemDetails
-            item={openDetailsItem}
+          <About
+            item={openAboutItem}
             onClose={() => {
-              setOpenDetailsProjectId(undefined);
+              setopenAbout(undefined);
             }}
           />
         </SettingsPanel>
